Ajout sécurité route DELETE

+++ controllers/sauces.js
@@ -46,15 +46,22 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
+  
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
+      if (!sauce) {
+        return res.status(404).json({error : new Error('Objet non trouvé!')})
+      }
+      if (sauce.userId !== req.auth.userId){
+        return res.status(401).json({error : new Error('Requète non autorisée')})
+      }else{
       const filename = sauce.imageUrl.split("/images/")[1];
       fs.unlink(`images/${filename}`, () => {
         Sauce.deleteOne({ _id: req.params.id })
           .then(() => res.status(200).json({ message: "Objet supprimé !" }))
           .catch((error) => res.status(400).json({ error }));
       });
-    })
+    }})
     .catch((error) => res.status(500).json({ error }));
 };
